refactor: extract batch helpers to _batch_helpers.py
I 6 helper di formattazione e silenziamento (format_years,
format_duration, print_table, build_row, silence_typer_echo,
silence_logger) erano in cmd_run.py, ingrossandolo di ~60 righe.

Estratti in toolkit/cli/_batch_helpers.py, importati da cmd_run.py.
Riduce cmd_run.py di ~60 righe, mantiene separazione netta.
